chore: ts cleanup
