+++ packages/vitest/src/api/setup.ts
@@ -7,7 +7,7 @@ import { WebSocketServer } from 'ws'
 import type { ViteDevServer } from 'vite'
 import type { File, TaskResultPack } from '@vitest/runner'
 import { API_PATH } from '../constants'
-import type { Vitest } from '../node'
+import type { Vitest } from '../node/core'
 import type { Awaitable, ModuleGraphData, UserConsoleLog } from '../types/general'
 import type { Reporter } from '../node/types/reporter'
 import { getModuleGraph, isPrimitive, noop, stringifyReplace } from '../utils'
