+++ packages/vitest/src/public/environments.ts
@@ -1,0 +1,7 @@
+export { environments as builtinEnvironments } from '../integrations/env/index'
+export { populateGlobal } from '../integrations/env/utils'
+export type {
+  EnvironmentReturn,
+  VmEnvironmentReturn,
+  Environment,
+} from '../types/environment'
