+++ packages/vitest/src/types/global.ts
@@ -1,9 +1,9 @@
 import type { Plugin as PrettyFormatPlugin } from '@vitest/pretty-format'
 import type { SnapshotState } from '@vitest/snapshot'
 import type { ExpectStatic, PromisifyAssertion, Tester } from '@vitest/expect'
-import type { UserConsoleLog } from './general'
 import type { VitestEnvironment } from '../node/types/config'
 import type { BenchmarkResult } from '../runtime/types/benchmark'
+import type { UserConsoleLog } from './general'
 
 declare global {
   // eslint-disable-next-line ts/no-namespace
