+++ packages/vitest/src/types/happy-dom-options.ts
