+++ tsconfig.base.json
@@ -19,9 +19,9 @@
       "@vitest/browser": ["./packages/browser/src/node/index.ts"],
       "@vitest/browser/client": ["./packages/browser/src/client/client.ts"],
       "~/*": ["./packages/ui/client/*"],
-      "vitest": ["./packages/vitest/src/index.ts"],
+      "vitest": ["./packages/vitest/src/public/index.ts"],
       "vitest/globals": ["./packages/vitest/globals.d.ts"],
-      "vitest/node": ["./packages/vitest/src/node/index.ts"],
+      "vitest/node": ["./packages/vitest/src/public/node.ts"],
       "vitest/execute": ["./packages/vitest/src/public/execute.ts"],
       "vitest/config": ["./packages/vitest/src/config.ts"],
       "vitest/coverage": ["./packages/vitest/src/coverage.ts"],
